moving cup retrieval up a layer
